requirements.txt
pytz
paddleocr
requests
paddlepaddle
Pillow
numpy